% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_components.R
\name{gg_facet}
\alias{gg_facet}
\title{GGPlot2 Facet Component}
\usage{
gg_facet(facet_formula = NULL, ...)
}
\arguments{
\item{facet_formula}{a valid r formula comprised of column names to create
facets. expects a string. formula can either be a '~rhs' formula for a
facet_wrap or 'lhs~rhs' to create a facet_grid. Default is NULL which
creates no facets}

\item{...}{}
}
\description{
Wrapper function for ggplot2 facet_wrap and facet_grid functions. Used
internally by sncr_* functions to handle facet arguements
}
\examples{
ggplot(mtcars, aes(x=mpg)) +
geom_histogram() +
gg_facet("~am")
}
\seealso{
ggplot2::facet_grid \url{http://ggplot2.tidyverse.org/reference/facet_grid.html}

ggplot2::facet_wrap \url{http://ggplot2.tidyverse.org/reference/facet_wrap.html}
}
