% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_interval_chart.R
\name{gg_interval_chart}
\alias{gg_interval_chart}
\title{GGplot Interval Chart}
\usage{
gg_interval_chart(df, x_variable, y_variable, color = sncr_pal()(1),
  alpha = 1, size = 1.1, shape = 18, chart_type = "pointrange",
  chart_fun = "mean_sdl", line = TRUE, lines_args = list(),
  facet_formula = NULL, facet_labeller = "both", facet_args = list(),
  coord = NULL, coord_args = list(), title = NULL, subtitle = NULL,
  x_axis_title = NULL, y_axis_title = NULL, caption = NULL,
  legend_args = list(), theme = "sncr", palette = "a2", ...)
}
\arguments{
\item{df}{data.frame object with variable to chart}

\item{x_variable}{x-variable to chart. expects a string. required input}

\item{y_variable}{y-variable to chart. expects a string. required input}

\item{color}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to color the outlines of
the histogram. expects a string. default is 'black'}

\item{alpha}{controls the opacity of the fill and color}

\item{size}{controls the line size}

\item{shape}{controls the shape of the point}

\item{chart_type}{chart types. accepts errorbar, linerange, pointrange, or
crossbar}

\item{chart_fun}{summary function to apply. accepts mean_cl_normal, mean_sdl,
mean, median, median_hilow, mean_se}

\item{line}{logical option to overlay a line between mean/median point values}

\item{lines_args}{list of arguments to pass to geom_line function. default is
an empty list}

\item{facet_formula}{a valid r formula comprised of column names to create
facets. expects a string. formula can either be a '~rhs' formula for a
facet_wrap or 'lhs~rhs' to create a facet_grid. Default is NULL which
creates no facets}

\item{facet_labeller}{takes a string. expecting name of one of label_*
functions. ex - "both" creates a facet label with both the column name and
the value. The two most common examples are 'both' and
'value'\url{http://ggplot2.tidyverse.org/reference/#section-facetting-labels}}

\item{facet_args}{a list of arguments to pass to either the facet_grid or
facet_wrap functions. default is a list with only he labeller parameter set
to both colname and field value.
\url{http://ggplot2.tidyverse.org/reference/facet_grid.html} or
\url{http://ggplot2.tidyverse.org/reference/facet_wrap.html}}

\item{coord}{name of coordinate transform. ex - 'flip' creates a coord_flip
and a 'trans' creates a coord_trans. expects a string. default is NULL with
creates no coordinate changes}

\item{coord_args}{list of arguments to pass to the coord_* function created.
default is empty list.}

\item{title}{chart title. expects a string}

\item{subtitle}{chart subtitle. expects a string}

\item{x_axis_title}{x-axis title. expects a string}

\item{y_axis_title}{y-axis title. expects a string}

\item{caption}{chart caption. expects a string}

\item{legend_args}{list of arguments to pass to guide_legend function to
control the legend formatting
\url{http://ggplot2.tidyverse.org/reference/guide_legend.html}. default is
empty list}

\item{theme}{argument to change the chart theme. expects a ggplot_theme name.
default is "sncr" which calls theme_sncr(). check out
\url{http://ggplot2.tidyverse.org/reference/#section-themes} for more
options}

\item{palette}{argument to change the color palette used is either the fill or
color is set to a variable. expects a palette name. default is "a2" for a2
palette}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{...}{Arguments passed on to \code{ggplot2::stat_summary}
\describe{
  \item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}
  \item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}
  \item{geom}{Use to override the default connection between
\code{geom_histogram}/\code{geom_freqpoly} and \code{stat_bin}.}
  \item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}
  \item{fun.data}{A function that is given the complete data and should
return a data frame with variables \code{ymin}, \code{y}, and \code{ymax}.}
  \item{fun.ymin}{Alternatively, supply three individual
functions that are each passed a vector of x's and should return a
single number.}
  \item{fun.y}{Alternatively, supply three individual
functions that are each passed a vector of x's and should return a
single number.}
  \item{fun.ymax}{Alternatively, supply three individual
functions that are each passed a vector of x's and should return a
single number.}
  \item{fun.args}{Optional additional arguments passed on to the functions.}
  \item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
  \item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
  \item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}}
}
\value{
returns a ggplot2 graphic object
}
\description{
This is a wrapper function to create a ggplot2 vertical interval chart. This
function unifies much of the ggplot2 functionality into a single function.
Default parameter values are set throughout to allow for quick interactive
charting, but is also highly customizable to allow for more sophisicated
charts.
}
\details{
This function takes any of the geom_errorbar arguments
\url{http://ggplot2.tidyverse.org/reference/geom_linerange.html}
}
\examples{
# Create a data set
library(dplyr)
d <- mtcars \%>\% mutate(am = as.factor(am), cyl = as.factor(cyl))

# Create series of basic charts
gg_interval_chart(d, y_variable = "mpg", x_variable = "am",
                  chart_type= "linerange" ,size=1.1)
gg_interval_chart(d, y_variable = "mpg", x_variable = "am",
                  chart_type= "pointrange", size=1.1, shape=18)
gg_interval_chart(d, y_variable = "mpg", x_variable = "am",
                  chart_type="errorbar", size=1.1, width = 0.2,
                  color='red',lines_args=list(color='blue'))
gg_interval_chart(d, y_variable = "mpg", x_variable = "am",
                  chart_type="crossbar", size=1.5, shape=18)
gg_interval_chart(d, y_variable = "mpg", x_variable = "am",
                  chart_type='errorbar',width=0.2,color='red')
gg_interval_chart(d, y_variable = "mpg", x_variable = "am",
                  chart_type='errorbar',width=0.2,
                  facet_formula='~cyl', color='cyl',
                  lines_args=list(color='blue'))
gg_interval_chart(d, y_variable = "mpg", x_variable = "am",
                  chart_type='errorbar', width=0.2, color='red',
                  lines_args=list(color='blue'))
}
\seealso{
geom_line
  \url{http://ggplot2.tidyverse.org/reference/geom_path.html}
}
