% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_components.R
\name{gg_label}
\alias{gg_label}
\title{GGPlot Label Component}
\usage{
gg_label(label = NULL, y = NULL, ...)
}
\arguments{
\item{label}{variable name or function such as ..count.. expects a string.
default is NULL. NULL value creates a blank geom}

\item{y}{variable name or function such as ..count.. xpects a string.
default is NULL}

\item{...}{additional arguments to geom_label}
}
\value{
returns either a geom_label or geom_blank object
}
\description{
Wrapper function to create a geom_label layer. The function takes two
arguements to the aes() function within the geom and any number of additional
arguments to the geom function
}
\seealso{
geom_label \url{http://ggplot2.tidyverse.org/reference/geom_text.html}
}
