% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_univariate_chart.R
\name{gg_univariate_chart}
\alias{gg_univariate_chart}
\title{GGplot2 Univariate Chart}
\usage{
gg_univariate_chart(df, x_variable, data_name = NULL, fill = "#1D3AB2",
  color = "grey15", alpha = 0.6, bins = 13, max_bars = 20,
  y_axis_title = "Count", caption = "Synchronoss", theme = "sncr",
  palette = "a2", plot_sizes = c(0.75, 0.25), ...)
}
\arguments{
\item{df}{data.frame object with variable to chart}

\item{x_variable}{x-variable to chart. expects a string. required input}

\item{data_name}{name for dataset. added to title if provided. default input
  is NULL which defers to dataset object name

  \code{\link{gg_bar_chart}} \code{\link{gg_boxplot}}
  \code{\link{gg_histogram}}}

\item{fill}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to fill the histogram
with color. expects a string. default is 'grey25'}

\item{color}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to color the outlines of
the histogram. expects a string. default is 'black'}

\item{alpha}{controls the opacity of the fill and color}

\item{bins}{number of bins used in histogram chart}

\item{max_bars}{maximum number of unique factor or character levels displayed
in chart}

\item{y_axis_title}{y-axis title. expects a string}

\item{caption}{chart caption. expects a string}

\item{theme}{argument to change the chart theme. expects a ggplot_theme name.
default is "sncr" which calls theme_sncr(). check out
\url{http://ggplot2.tidyverse.org/reference/#section-themes} for more
options}

\item{palette}{argument to change the color palette used is either the fill or
color is set to a variable. expects a palette name. default is "a2" for a2
palette}

\item{plot_sizes}{vector of two ratios that add to 1. Sets the size ratios of
the two charts.}
}
\value{
returns a ggplot2 object as a result of a call to a gg_* charting
  function
}
\description{
GGplot2 Univariate Chart
}
\examples{

gg_univariate_chart(mtcars, "mpg")
gg_univariate_chart(mutate(mtcars, am = as.factor(am)), "am")
}
