% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-palette-theme.R
\name{sncr_pal}
\alias{sncr_pal}
\title{Return function to interpolate a synchronoss color palette}
\usage{
sncr_pal(palette = "a2", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in sncr_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a synchronoss color palette
}
