% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-palette-theme.R
\name{theme_sncr}
\alias{theme_sncr}
\title{Synchronoss GGplot2 Theme}
\usage{
theme_sncr(base_size = 12, base_family = "serif")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggtheme
}
\description{
Uses theme_minimal as base and replaces some panel, plot, axis, legend, and
panel.grib properties
}
